+++ sdk/rollup.config.js
@@ -18,6 +18,7 @@ export default [
             'physics3d/index': 'src/physics3d/index.ts',
             'douyin/index': 'src/platform/douyin/index.ts',
             'spine/index': 'src/spine/index.ts',
+            'tilemap/index': 'src/tilemap/index.ts',
             'dragonbones/index': 'src/dragonbones/index.ts',
             'wasm': 'src/wasm.ts',
         },